Fixed race conditions in Child and Pool classes
The Child class' wait method checks the isRunning method to determine whether the child has terminated or not.  However, a race condition exists where the process has terminated before the wait method is called.  In this case, isRunning returns false, and then the wait method will return false, rather than calling waitpid to reap the process and get the status.  Instead of checking isRunning, the method now checks to see if the status has been set in the instance already.  If not, then the wait method hasn't been previously called for this child and so it calls the waitpid function.

In Pool class, it checks if the child process is running by again calling its isRunning method.  If the child process terminates before the wait method of Pool is called, then again the Child object's wait method is never called.  The test for whether isRunning is false for the Child object is omitted, so that it must call the wait method of the Child object.  If the child has already had its wait method called, then it will just return false as intended.